Updating project structure
Class Program divided into two more
classes for better code readability
